Load config before page is rendered
This reduces the risk that conditionsFailed is triggered because the session will be
established in time.
